move components of task screen to separate package separate task screen from its content
