+++ app/src/main/java/com/baghdad/tudee/ui/screens/tasks/TasksInteractionListener.kt
@@ -3,7 +3,7 @@ package com.baghdad.tudee.ui.screens.tasks
 import com.baghdad.tudee.domain.entity.Task
 import kotlinx.datetime.LocalDate
 
-interface TasksInteractionHandler {
+interface TasksInteractionListener {
     fun onTabSelected(selectedTab: Task.State)
     fun onDateSelectedFromHorizontalRow(selectedDate: LocalDate)
     fun onDatePickedFromDateDialog(selectedDate: LocalDate)
