+++ app/src/main/java/com/baghdad/tudee/ui/screens/tasks/components/SwipeToDeleteCard.kt
@@ -1,4 +1,4 @@
-package com.baghdad.tudee.ui.screens.tasks
+package com.baghdad.tudee.ui.screens.tasks.components
 
 import androidx.compose.runtime.Composable
