+++ app/src/main/java/com/baghdad/tudee/ui/screens/tasks/components/TasksHeader.kt
@@ -1,4 +1,4 @@
-package com.baghdad.tudee.ui.screens.tasks
+package com.baghdad.tudee.ui.screens.tasks.components
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.ExperimentalAnimationApi
@@ -6,7 +6,6 @@ import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.with
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
